requirements.txt
# NEXUS — samlet requirements (Jordan + NEXUS)

# ── Core AI ───────────────────────────────────────────────────────────────────
anthropic>=0.40.0
python-dotenv>=1.0.0

# ── Telegram ──────────────────────────────────────────────────────────────────
python-telegram-bot>=21.0

# ── HTTP / scraping ───────────────────────────────────────────────────────────
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ── Data ─────────────────────────────────────────────────────────────────────
pydantic>=2.0.0

# ── Scheduler ────────────────────────────────────────────────────────────────
apscheduler>=3.10.0
pytz>=2024.1

# ── Web API ───────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn>=0.30.0

# ── Browser automation ────────────────────────────────────────────────────────
playwright>=1.44.0

# ── Voice / speech ───────────────────────────────────────────────────────────
faster-whisper>=1.0.0
elevenlabs>=1.0.0

# ── Google ────────────────────────────────────────────────────────────────────
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0
google-generativeai>=0.7.0

# ── Social / integrasjoner ────────────────────────────────────────────────────
PyGithub>=2.3.0
stripe>=9.0.0
instagrapi>=2.0.0

# ── Email ─────────────────────────────────────────────────────────────────────
resend>=0.8.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── PDF ───────────────────────────────────────────────────────────────────────
pypdf>=4.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
